fixing custom repos browser title
